Update video constraints to use standard resolutions
The existing ideal resolution was set to a non-standard 360x240 - change
it to match the QVGA standard of 320x240.

The existing max resolution was set to a very low value that may not be
supported by some webcams. Increase this to the VGA standard to open
support for a broader selection of webcams.

ideal: 320x240 (QVGA)
min: 160x120 (QQVGA)
max: 640x480 (VGA)