Allow connection to fall back to audio-only when video isn't available
When video isn't available either due to the lack of a webcam or because
the user denied video permissions, the initialization of the stream will
fail. If the returned stream is null, FVTT will automatically attempt to
fall back to an audio-only stream.

Returning in this way also allows FVTT to properly display the error to
the user within game.