audio: Fix audio-only input
assignStreamToVideo should be used to assign a stream to the video
element even if no video tracks exist. The video element is used even
when a stream is audio-only.

Fix reversed logic on canBroadcastAudio check to properly set up the
audio elements. Also always call setVideoStream in case an audio track
is updated without a video track.