refactor: Remove a lot of prototype overrides to closer match FVTT
FVTT's WebRTC implementation was build around a lot of expectations from
the default EasyRTC service. To better align with those expectations,
refactor the code to work the way Foundry is expecting.

The allows some of the code to be reduced and will hopefully provide
more stability and easier debugging in the future.