security: Kick non FVTT users if the standalone user option is disabled
If "Allow standalone Jitsi users" is not enabled and a user enters the
meeting outside of FVTT, they wouldn't be shown to other users but would
be able to see & hear everything in the meeting. For security, kick any
user who doesn't have a matching FVTT ID unless this option is enabled.