onSettingsChanged: Hack fix for changing audio and video input devices
When attempting to fully implement closeLocalStream and initLocalStream,
there's an issue where the microphone volume is getting doubled when the
audio input device is changed. To avoid this, but still allow for the
change to take effect, force a full page reload when this setting is
changed.

TODO: Fix this properly after other RTC changes are in place.