jitsirtc: Properly handle server URL between custom and default server
We need to read the configured server type so we only set the URL to the
default when custom isn't being used. As it was, if a value was set in
the server settings, it would be used even if the setting was no longer
configured to use that custom server.