jitsirtc: Properly handle missing audio and video devices
We need to cleanly handle either no camera or microphone being detected
as well as a device that is no longer connected being set as the device
to use.

If no devices are available, or if a missing device is selected, set the
device to null so the connection can still establish as audio or video
only.