jitsi: Set the FVTT user ID as a property instead of displayName
Using the displayName as the value holder for the FVTT User ID makes
external jitsi user's have a bad UX. Instead, save the ID as a property
and set the display name properly as the FVTT user name.

Since we can't set the property value with external Jitsi Meet users,
set the ID as a formatted part of the name and parse it out in the case
of externally joining users.